partly finishing this part, now it can read json files
