+++ brutal_force/README.md
