+++ brutal_force/jsonfile/topology.json
@@ -1,0 +1,24 @@
+{
+    "topology": [
+        {
+            "src": 1,
+            "dst": 3
+        },
+        {
+            "src": 2,
+            "dst": 6
+        },
+        {
+            "src": 4,
+            "dst": 7
+        },
+        {
+            "src": 5,
+            "dst": 10
+        },
+        {
+            "src": 8,
+            "dst": 9
+        }
+    ]
+}
